Revert "Removed Old Versions of Files."
This reverts commit e8c130dd9e385a75200ec1c4c664d36051e70bed, reversing
changes made to 5fcf62777cf84b66d01783a550542901e8c8aa37.

+++ Position.java
@@ -9,38 +9,38 @@
  * Position is a class which defines a datatype Position which contains an x and y coord.
  *
  */
-//change doubles to floats when implementing smooth movement.
+//change ints to floats when implementing smooth movement.
 public class Position {
 	/**
 	 * the x coord for a position.
 	 * the y coord for a position.
 	 */
-	private double positionX;
-	private double positionY;
+	private int positionX;
+	private int positionY;
 	
 	/**
 	 * This defines a position of x and y.
 	 * @param x
 	 * @param y
 	 */
-	public Position (double x, double y) {
+	public Position (int x, int y) {
 		this.positionX = x;
 		this.positionY = y;
 	}
 	
 	/**
-	 * Returns the x and y coord as an array of doubleegers.
+	 * Returns the x and y coord as an array of integers.
 	 *	@return 1D array consisting of x and y.
 	 */
-	public double[] getPosition () {
-		double retrunArray[] = {positionX, positionY};
+	public int[] getPosition () {
+		int retrunArray[] = {positionX, positionY};
 		return retrunArray;
 	}
 	
 	/**
 	 * Overwrites the current x and y of a position,
 	 */
-	public void setPosition (double x, double y) {
+	public void setPosition (int x, int y) {
 		this.positionX = x;
 		this.positionY = y;
 	}
